fix: type RinseFlush.time_seconds_step as float
The API returns fractional step sizes (e.g. timeSecondsStep: 0.1), but
the field was typed as int, truncating the value to 0 and losing the
supported step size. Type it as float and refresh the affected
snapshots.

Addresses review comment on RinseFlush.time_seconds_step precision.

+++ pylamarzocco/models/_config.py
@@ -368,7 +368,7 @@ class RinseFlush(BaseWidgetOutput):
     time_seconds: float = field(metadata=field_options(alias="timeSeconds"))
     time_seconds_min: int = field(metadata=field_options(alias="timeSecondsMin"))
     time_seconds_max: int = field(metadata=field_options(alias="timeSecondsMax"))
-    time_seconds_step: int = field(metadata=field_options(alias="timeSecondsStep"))
+    time_seconds_step: float = field(metadata=field_options(alias="timeSecondsStep"))
 
 
 @dataclass(kw_only=True)
